录音转写客户端相关代码格式调整
