+++ websdk-java-speech/src/main/java/cn/xfyun/api/LfasrClient.java
@@ -2,291 +2,483 @@ package cn.xfyun.api;
 
 import cn.xfyun.service.lfasr.task.*;
 import cn.xfyun.exception.LfasrException;
-import cn.xfyun.model.response.lfasr.LfasrMessage;
+import cn.xfyun.model.response.lfasr.LfasrResponse;
 import cn.xfyun.model.sign.LfasrSignature;
 import cn.xfyun.service.lfasr.LfasrExecutorService;
-import cn.xfyun.util.SliceIdGenerator;
 import cn.xfyun.util.StringUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.io.File;
-import java.io.FileInputStream;
+import java.net.URLEncoder;
 import java.security.SignatureException;
-import java.util.Arrays;
 import java.util.HashMap;
 import java.util.Map;
 
 /**
- * 转写client
- *
- * @author : iflytek
- * @date : 2021年03月15日
+ * 录音文件转写客户端
  */
 public class LfasrClient {
+
     private static final Logger logger = LoggerFactory.getLogger(LfasrClient.class);
 
     /**
-     * 最大500M
+     * 文件限制最大500M
      */
     private static final int FILE_UPLOAD_MAXSIZE = 524288000;
+
     /**
-     * 建议10M一片
+     * 应用ID（必填）
      */
-    private static int SLICE_SIZE = 10485760;
-    private final String secretKey;
-    private volatile LfasrExecutorService lfasrExecutorService;
     private String appId;
+    
     /**
-     * 文件名称（带后缀）
+     * 应用密钥（必填）
      */
-    private String file_name;
+    private String secretKey;
+    
+    /**
+     * 转写执行服务
+     */
+    private volatile LfasrExecutorService lfasrExecutorService;
+    
+    /**
+     * 音频文件名称，最好携带音频真实的后缀名，避免影响转码（必填）
+     */
+    private String fileName;
+    
+    /**
+     * 音频文件大小（字节数）（必填）
+     * 当前只针对本地文件流方式校验，使用url外链方式不校验，可随机传一个数字
+     */
+    private Long fileSize;
+    
+    /**
+     * 音频真实时长，当前未验证，可随机传一个数字（必填）
+     */
+    private Long duration;
 
     /**
-     * 文件分片大小（建议分片大小为10M）
-     */
-    private Integer slice_size;
-
-    /**
-     * 文件分片数目（建议分片大小为10M，若文件<10M，则slice_num=1）
-     */
-    private Integer slice_num;
-
-    /**
-     * 转写类型，默认 0
-     * 0: (标准版，格式: wav,flac,opus,mp3,m4a)
-     * 2: (电话版，已取消)
-     */
-    private String lfasr_type;
-
-    /**
-     * 转写结果是否包含分词信息
-     */
-    private String has_participle;
-
-    /**
-     * 转写结果中最大的候选词个数
-     */
-    private String max_alternatives;
-    /**
-     * 发音人个数，可选值：0-10，0表示盲分
-     * 注：发音人分离目前还是测试效果达不到商用标准，如测试无法满足您的需求，请慎用该功能。
-     */
-    private String speaker_number;
-    /**
-     * 转写结果中是否包含发音人分离信息
-     */
-    private String has_seperate;
-
-    /**
-     * 支持两种参数
-     * 1: 通用角色分离
-     * 2: 电话信道角色分离（适用于speaker_number为2的说话场景）
-     */
-    private String role_type;
-
-    /**
-     * 语种
-     * cn:中英文&中文（默认）
-     * en:英文（英文不支持热词）
+     * 语种类型（非必填）
+     * cn：中文，通用方言（包括普通话、天津、河北、东北、甘肃、山东、太原）（默认）
+     * en：英文
+     * ja：日语
+     * ko：韩语
+     * ru：俄语
+     * fr：法语
+     * es：西班牙语
+     * vi：越南语
+     * ar：阿拉伯语
+     * cn_xinanese：西南官话（包括四川、重庆、云南、贵州）
+     * cn_cantonese：粤语
+     * cn_henanese： 河南
+     * cn_uyghur：维吾尔语
+     * cn_tibetan：藏语
+     * ar:阿拉伯语
+     * de:德语
+     * it:意大利语
      */
     private String language;
-
+    
     /**
-     * 垂直领域个性化参数:
-     * 法院: court
-     * 教育: edu
-     * 金融: finance
-     * 医疗: medical
-     * 科技: tech
+     * 回调地址（非必填）
+     * 订单完成时回调该地址通知完成支持get请求，长度限制512
+     * 参数：orderId为订单号、status为订单状态: 1(转写识别成功)、-1(转写识别失败)
+     */
+    private String callbackUrl;
+    
+    /**
+     * 热词，用以提升专业词汇的识别率（非必填）
+     * 格式：热词1|热词2|热词3
+     * 单个热词长度：[2,16]，热词个数限制200个
+     */
+    private String hotWord;
+    
+    /**
+     * 多候选开关（非必填）
+     * 0：关闭(默认)
+     * 1：打开
+     */
+    private Short candidate;
+    
+    /**
+     * 是否开启角色分离（非必填）
+     * 0：不开启角色分离(默认)
+     * 1：通用角色分离
+     */
+    private Short roleType;
+    
+    /**
+     * 说话人数，取值范围0-10，默认为0进行盲分（非必填）
+     * 注：该字段只有在开通了角色分离功能的前提下才会生效
+     */
+    private Short roleNum;
+    
+    /**
+     * 领域个性化参数（非必填）
+     * court：法律
+     * edu：教育
+     * finance：金融
+     * medical：医疗
+     * tech：科技
+     * culture：人文历史
+     * isp：运营商
+     * sport：体育
+     * gov：政府
+     * game：游戏
+     * ecom：电商
+     * mil：军事
+     * com：企业
+     * life：生活
+     * ent：娱乐
+     * car：汽车
      */
     private String pd;
+    
+    /**
+     * 转写音频上传方式（非必填）
+     * fileStream：文件流(默认)
+     * urlLink：音频url外链
+     */
+    private String audioMode;
+    
+    /**
+     * 音频url外链地址（非必填，audioMode为urlLink时必填）
+     * 当audioMode为urlLink时该值必传
+     * 如果url中包含特殊字符，audioUrl需要UrlEncode，长度限制512
+     */
+    private String audioUrl;
+    
+    /**
+     * 是否标准pcm/wav(16k/16bit/单声道)（非必填）
+     * 0：非标准wav(默认)
+     * 1：标准pcm/wav
+     */
+    private Integer standardWav;
+    
+    /**
+     * 语言识别模式选择（非必填）
+     * language为cn时：
+     * 1：自动中英文模式(默认)
+     * 2：中文模式（可能包含少量英文）
+     * 4：纯中文模式（不包含英文）
+     */
+    private Integer languageType;
+    
+    /**
+     * 按声道分轨转写模式（支持语种：cn、en）（非必填）
+     * 1：不分轨模式(默认)
+     * 2：双声道分轨模式
+     * 备注：如果转写任务使用双声道分轨模式，角色分离(roleType)功能失效
+     */
+    private Short trackMode;
+    
+    /**
+     * 需要翻译的语种(转写语种和翻译语种不能相同)（非必填）
+     */
+    private String transLanguage;
+    
+    /**
+     * 翻译模式（默认2：按段落进行翻译，目前只支持按段落进行翻译）（非必填）
+     * 1：按VAD进行翻译
+     * 2：按段落进行翻译
+     * 3：按整篇进行翻译
+     */
+    private Short transMode;
+    
+    /**
+     * 控制分段的最大字数，取值范围[0-500]，不传使用引擎默认值（非必填）
+     */
+    private Integer engSegMax;
+    
+    /**
+     * 控制分段的最小字数，取值范围[0-50]，不传使用引擎默认值（非必填）
+     */
+    private Integer engSegMin;
+    
+    /**
+     * 控制分段字数的权重，权重比越高，表示引擎分段逻辑采用字数控制分段的比重越高（非必填）
+     * 取值(0-0.05)不传即不采用字数控制分段，采用引擎默认分段逻辑
+     */
+    private Float engSegWeight;
+    
+    /**
+     * 顺滑开关（非必填）
+     * true：表示开启(默认)
+     * false：表示关闭
+     */
+    private Boolean engSmoothproc;
+    
+    /**
+     * 口语规整开关，口语规整是顺滑的升级版本（非必填）
+     * true：表示开启
+     * false：表示关闭(默认)
+     */
+    private Boolean engColloqproc;
+    
+    /**
+     * 远近场模式（非必填）
+     * 1：远场模式(默认)
+     * 2：近场模式
+     */
+    private Integer engVadMdn;
+    
+    /**
+     * 首尾是否带静音信息（非必填）
+     * 0：不显示
+     * 1：显示(默认)
+     */
+    private Integer engVadMargin;
+    
+    /**
+     * 针对粤语转写后的字体转换（非必填）
+     * 0：输出简体
+     * 1：输出繁体(默认)
+     */
+    private Integer engRlang;
 
     /**
-     * builder 字段赋值
+     * 通过Builder构建LfasrClient实例
      *
-     * @param builder
+     * @param builder 构建器
      */
     public LfasrClient(LfasrClient.Builder builder) {
+
+
+
         this.appId = builder.appId;
         this.secretKey = builder.secretKey;
-        this.file_name = builder.file_name;
-        this.slice_size = builder.slice_size;
-        this.lfasr_type = builder.lfasr_type;
-        this.has_participle = builder.has_participle;
-        this.max_alternatives = builder.max_alternatives;
-        this.speaker_number = builder.speaker_number;
-        this.has_seperate = builder.has_seperate;
-        this.role_type = builder.role_type;
+        this.fileName = builder.fileName;
+        this.fileSize = builder.fileSize;
+        this.duration = builder.duration;
         this.language = builder.language;
+        this.callbackUrl = builder.callbackUrl;
+        this.hotWord = builder.hotWord;
+        this.candidate = builder.candidate;
+        this.roleType = builder.roleType;
+        this.roleNum = builder.roleNum;
         this.pd = builder.pd;
+        this.audioMode = builder.audioMode;
+        this.audioUrl = builder.audioUrl;
+        this.standardWav = builder.standardWav;
+        this.languageType = builder.languageType;
+        this.trackMode = builder.trackMode;
+        this.transLanguage = builder.transLanguage;
+        this.transMode = builder.transMode;
+        this.engSegMax = builder.engSegMax;
+        this.engSegMin = builder.engSegMin;
+        this.engSegWeight = builder.engSegWeight;
+        this.engSmoothproc = builder.engSmoothproc;
+        this.engColloqproc = builder.engColloqproc;
+        this.engVadMdn = builder.engVadMdn;
+        this.engVadMargin = builder.engVadMargin;
+        this.engRlang = builder.engRlang;
     }
 
     /**
-     * 预处理接口
-     * 上传待转写音频文件的基本信息（文件名、大小）和分片信息（建议分片大小设置为10M，若无需分片，slice_num=1）和相关的可配置参数。
-     * 调用成功，返回任务ID（task_id，转写任务的唯一标识），是后续接口的必传参数。
-     *
-     * @param audioFilePath
-     * @return
-     * @throws LfasrException
-     * @throws SignatureException
+     * 本地音频文件上传接口
+     * 
+     * @param audioFilePath 音频文件路径
+     * @return 调用成功时返回orderId（转写任务的唯一标识），是后续接口的必传参数。
+     * @throws LfasrException 转写异常
+     * @throws SignatureException 签名异常
      */
-    public LfasrMessage prepare(String audioFilePath) throws LfasrException, SignatureException {
+    public LfasrResponse uploadFile(String audioFilePath) throws LfasrException, SignatureException {
         // 业务校验
         if (StringUtils.isNullOrEmpty(audioFilePath)) {
-            throw new LfasrException("audioFilePath is null!");
+            throw new LfasrException("音频文件地址为空!");
         }
-        File audio = new File(audioFilePath);
-        if (!audio.exists()) {
-            throw new LfasrException(audioFilePath + " is not exists!");
+        
+        // 文件存在性和大小校验
+        File audioFile = new File(audioFilePath);
+        if (!audioFile.exists()) {
+            throw new LfasrException(audioFilePath + "不存在!");
         }
-        long length = audio.length();
-        if (length > FILE_UPLOAD_MAXSIZE) {
-            throw new LfasrException(audioFilePath + " is too large! (500M)");
+        if (audioFile.length() > FILE_UPLOAD_MAXSIZE) {
+            throw new LfasrException(audioFilePath + " 文件过大! (500M)");
         }
+        
+        // 构建参数
         Map<String, String> param = new HashMap<>(32);
-        paramHandler(param, length, audio);
-        return this.lfasrExecutorService.exec(new PrepareTask(new LfasrSignature(appId, secretKey), param));
+        paramHandler(param, audioFile, null);
+        
+        // 执行文件任务
+        return this.lfasrExecutorService.exec(new UploadFileTask(new LfasrSignature(appId, secretKey), param, audioFile));
     }
 
     /**
-     * 文件分片上传接口
-     * 按预处理设置的分片信息（slice_num）依次上传音频切片（文件以二进制方式读取上传），
-     * 直到全部切片上传成功（如预处理时 slice_num=2，则需将音频切分成两部分，slice_id=aaaaaaaaaa和aaaaaaaaab，并按顺序调用该接口）；
-     * 上一切片成功上传，才可进行下一切片的上传操作。调用过程中若出现异常，可重试若干次。
-     *
-     * @param taskId
-     * @param audioFilePath
-     * @throws LfasrException
+     * 音频URL上传接口
+     * 
+     * @param audioUrl 音频链接
+     * @return 调用成功时返回orderId（转写任务的唯一标识），是后续接口的必传参数。
+     * @throws LfasrException 转写异常
+     * @throws SignatureException 签名异常
      */
-    public void uploadFile(String taskId, String audioFilePath) throws LfasrException {
-        File audio = new File(audioFilePath);
-        try (FileInputStream fis = new FileInputStream(audio)) {
-            byte[] slice = new byte[SLICE_SIZE];
-            SliceIdGenerator generator = new SliceIdGenerator();
-            int len;
-            while ((len = fis.read(slice)) > 0) {
-                byte[] data = Arrays.copyOfRange(slice, 0, len);
-                UploadTask uploadTask = new UploadTask(new LfasrSignature(appId, secretKey), taskId, generator.getNextSliceId(), data);
-                this.lfasrExecutorService.exec(uploadTask);
-            }
-        } catch (Exception e) {
-            logger.error(audioFilePath + "上传处理失败", e);
-            throw new LfasrException("转写上传文件读取错误");
+    public LfasrResponse uploadUrl(String audioUrl) throws LfasrException, SignatureException {
+
+        if (StringUtils.isNullOrEmpty(audioUrl)) {
+            throw new LfasrException("音频链接为空!");
         }
-    }
-
-    /**
-     * 合并文件接口
-     * 全部文件切片上传成功后，调用该接口，通知服务端进行文件合并与转写操作。
-     * 该接口不会返回转写结果，而是通知服务端将任务列入转写计划。转写的结果通过 getResult 接口获取。
-     *
-     * @param taskId
-     * @return
-     * @throws LfasrException
-     * @throws SignatureException
-     */
-    public LfasrMessage mergeFile(String taskId) throws LfasrException, SignatureException {
-        return this.lfasrExecutorService.exec(new MergeTask(new LfasrSignature(appId, secretKey), taskId));
-    }
-
-    /**
-     * 文件上传接口
-     * 包含：
-     * 1、prepare
-     * 2、uploadFile
-     * 3、merge
-     *
-     * @param audioFilePath
-     * @return
-     * @throws LfasrException
-     * @throws SignatureException
-     */
-    public LfasrMessage upload(String audioFilePath) throws LfasrException, SignatureException {
-        LfasrMessage message = prepare(audioFilePath);
-        String taskId = message.getData();
-        logger.info("taskId: " + taskId);
-        uploadFile(taskId, audioFilePath);
-        mergeFile(taskId);
-        return message;
-    }
-
-    /**
-     * 查询处理进度接口
-     * 在调用方发出合并文件请求后，服务端已将任务列入计划。在获取结果前，调用方需轮询该接口查询任务当前状态。
-     * 当且仅当任务状态=9（转写结果上传完成），才可调用获取结果接口获取转写结果。
-     * 轮询策略由调用方决定，建议每隔10分钟轮询一次。状态码说明见附录。
-     *
-     * @param taskId
-     * @return
-     * @throws SignatureException
-     */
-    public LfasrMessage getProgress(String taskId) throws SignatureException {
-        return this.lfasrExecutorService.exec(new QueryProgressTask(new LfasrSignature(appId, secretKey), taskId));
+        
+        // 构建参数
+        this.audioMode = "urlLink";
+        this.audioUrl = audioUrl;
+        this.fileName = StringUtils.isNullOrEmpty(this.fileName) ? audioUrl.substring(audioUrl.lastIndexOf("/") + 1) : this.fileName;
+        this.fileSize = this.fileSize != null ? this.fileSize : 1L;
+        Map<String, String> param = new HashMap<>(32);
+        paramHandler(param, null, audioUrl);
+        
+        // 执行文件任务
+        return this.lfasrExecutorService.exec(new UploadUrlTask(new LfasrSignature(appId, secretKey), param, audioUrl));
     }
 
     /**
      * 获取结果接口
-     * 当任务处理进度状态=9（见查询处理进度接口），调用该接口获取转写结果。这是转写流程的最后一步。
      *
-     * @param taskId
-     * @return
-     * @throws SignatureException
+     * @param orderId 转写订单ID
+     * @return 转写结果
+     * @throws SignatureException 签名异常
      */
-    public LfasrMessage getResult(String taskId) throws SignatureException {
-        return this.lfasrExecutorService.exec(new PullResultTask(new LfasrSignature(appId, secretKey), taskId));
+    public LfasrResponse getResult(String orderId) throws SignatureException {
+        return getResult(orderId, null);
+    }
+
+    /**
+     * 获取结果接口
+     *
+     * @param orderId 转写订单ID
+     * @param resultType 查询结果类型
+     * - 转写结果：transfer（默认）
+     * - 翻译结果：translate
+     * - 质检结果：predict
+     * - 组合结果查询：多个类型结果使用”,”隔开，目前只支持转写和质检结果一起返回，不支持转写和翻译结果一起返回（如果任务有失败则只返回处理成功的结果）
+     * - 转写和质检结果组合返回：transfer,predict
+     * - 注：使用质检功能请先在控制台开启
+     * @return 转写结果
+     * @throws SignatureException 签名异常
+     */
+    public LfasrResponse getResult(String orderId, String resultType) throws SignatureException {
+        return this.lfasrExecutorService.exec(new PullResultTask(new LfasrSignature(appId, secretKey), orderId, resultType));
     }
 
     /**
      * 接口参数处理
+     * 处理所有请求参数并添加到参数Map中
      *
-     * @param param
-     * @param length
-     * @param audio
+     * @param param 参数Map
+     * @param length 文件长度
+     * @param audio 音频文件
      */
-    private void paramHandler(Map<String, String> param, long length, File audio) {
-        LfasrClient.SLICE_SIZE = this.slice_size == null ? SLICE_SIZE : this.slice_size;
+    private void paramHandler(Map<String, String> param, File audioFile, String audioUrl) {
+        // 必填参数设置
+        param.put("fileName", StringUtils.isNullOrEmpty(this.fileName) ? audioFile.getName() : this.fileName);
+        param.put("fileSize", this.fileSize == null ? audioFile.length() + "" : this.fileSize + "");
+        param.put("duration", this.duration != null ? this.duration + "" : "1");
 
-        long sliceNum = length / SLICE_SIZE + ((length % SLICE_SIZE == 0L) ? 0L : 1L);
-
-        param.put("file_len", length + "");
-        param.put("file_name", StringUtils.isNullOrEmpty(this.file_name) ? audio.getName() : this.file_name);
-        param.put("slice_num", sliceNum + "");
-
-        if (!StringUtils.isNullOrEmpty(this.lfasr_type)) {
-            param.put("lfasr_type", this.lfasr_type);
-        }
-
-        if (!StringUtils.isNullOrEmpty(this.has_participle)) {
-            param.put("has_participle", this.has_participle);
-        }
-
-        if (!StringUtils.isNullOrEmpty(this.max_alternatives)) {
-            param.put("max_alternatives", this.max_alternatives);
-        }
-        if (!StringUtils.isNullOrEmpty(this.has_seperate)) {
-            param.put("has_seperate", this.has_seperate);
-        }
-
-        if (!StringUtils.isNullOrEmpty(this.speaker_number)) {
-            param.put("speaker_number", this.speaker_number);
-        }
-
-        if (!StringUtils.isNullOrEmpty(this.role_type)) {
-            param.put("role_type", this.role_type);
+        if ("urlLink".equals(this.audioMode)) {
+            if (StringUtils.isNullOrEmpty(this.audioUrl)) {
+                throw new LfasrException("音频链接为空!");
+            }
+            try {
+                String encodedUrl = URLEncoder.encode(this.audioUrl, "UTF-8");
+                param.put("audioUrl", encodedUrl);
+            } catch (java.io.UnsupportedEncodingException e) {
+                throw new LfasrException("URL编码失败: " + e.getMessage());
+            }
+            param.put("audioUrl", this.audioUrl);
         }
 
         if (!StringUtils.isNullOrEmpty(this.language)) {
             param.put("language", this.language);
         }
+        
+        if (!StringUtils.isNullOrEmpty(this.callbackUrl)) {
+            param.put("callbackUrl", this.callbackUrl);
+        }
+       
+        if (!StringUtils.isNullOrEmpty(this.hotWord)) {
+            param.put("hot_word", this.hotWord);
+        }
+
+        if (this.candidate != null) {
+            param.put("candidate", this.candidate + "");
+        }
+        
+        if (this.roleType != null) {
+            param.put("roleType", this.roleType + "");
+        }
+
+        if (this.roleNum != null) {
+            param.put("roleNum", this.roleNum + "");
+        }
 
         if (!StringUtils.isNullOrEmpty(this.pd)) {
             param.put("pd", this.pd);
         }
+   
+        if (!StringUtils.isNullOrEmpty(this.audioMode)) {
+            param.put("audioMode", this.audioMode);
+        }
+
+        if (!StringUtils.isNullOrEmpty(this.audioUrl)) {
+            param.put("audioUrl", this.audioUrl);
+        }
+
+        if (this.standardWav != null) {
+            param.put("standardWav", this.standardWav + "");
+        }
+        
+        if (this.languageType != null) {
+            param.put("languageType", this.languageType + "");
+        }
+        
+        if (this.trackMode != null) {
+            param.put("trackMode", this.trackMode + "");
+        }
+        
+        if (!StringUtils.isNullOrEmpty(this.transLanguage)) {
+            param.put("transLanguage", this.transLanguage);
+        }
+        
+        if (this.transMode != null) {
+            param.put("transMode", this.transMode + "");
+        }
+        
+        if (this.engSegMax != null) {
+            param.put("eng_seg_max", this.engSegMax + "");
+        }
+        
+        if (this.engSegMin != null) {
+            param.put("eng_seg_min", this.engSegMin + "");
+        }
+        
+        if (this.engSegWeight != null) {
+            param.put("eng_seg_weight", this.engSegWeight + "");
+        }
+        
+        if (this.engSmoothproc != null) {
+            param.put("eng_smoothproc", this.engSmoothproc ? "true" : "false");
+        }
+        
+        if (this.engColloqproc != null) {
+            param.put("eng_colloqproc", this.engColloqproc ? "true" : "false");
+        }
+        
+        if (this.engVadMdn != null) {
+            param.put("eng_vad_mdn", this.engVadMdn + "");
+        }
+        
+        if (this.engVadMargin != null) {
+            param.put("eng_vad_margin", this.engVadMargin + "");
+        }
+        
+        if (this.engRlang != null) {
+            param.put("eng_rlang", this.engRlang + "");
+        }
     }
 
     public static final class Builder {
+
         private final String appId;
         private final String secretKey;
         private Integer coreThreads = 10;
@@ -295,26 +487,31 @@ public class LfasrClient {
         private Integer connTimeout = 10000;
         private Integer soTimeout = 30000;
 
-        private String file_name;
-
-        private Integer slice_size;
-
-        private String lfasr_type;
-
-        private String has_participle;
-
-        private String max_alternatives;
-
-        private String speaker_number;
-
-        private String has_seperate;
-
-        private String role_type;
-
+        private String fileName;
+        private Long fileSize;
+        private Long duration;
         private String language;
-
+        private String callbackUrl;
+        private String hotWord;
+        private Short candidate;
+        private Short roleType;
+        private Short roleNum;
         private String pd;
-
+        private String audioMode;
+        private String audioUrl;
+        private Integer standardWav;
+        private Integer languageType;
+        private Short trackMode;
+        private String transLanguage;
+        private Short transMode;
+        private Integer engSegMax;
+        private Integer engSegMin;
+        private Float engSegWeight;
+        private Boolean engSmoothproc;
+        private Boolean engColloqproc;
+        private Integer engVadMdn;
+        private Integer engVadMargin;
+        private Integer engRlang;
 
         public Builder(String appId, String secretKey) {
             this.appId = appId;
@@ -346,44 +543,18 @@ public class LfasrClient {
             return this;
         }
 
-        public LfasrClient.Builder file_name(String file_name) {
-            this.file_name = file_name;
+        public LfasrClient.Builder fileName(String fileName) {
+            this.fileName = fileName;
             return this;
         }
 
-        public LfasrClient.Builder slice(Integer slice_size) {
-            this.slice_size = slice_size;
+        public LfasrClient.Builder fileSize(Long fileSize) {
+            this.fileSize = fileSize;
             return this;
         }
 
-        public LfasrClient.Builder lfasr_type(String lfasr_type) {
-            this.lfasr_type = lfasr_type;
-            return this;
-        }
-
-        public LfasrClient.Builder has_participle(String has_participle) {
-            this.has_participle = has_participle;
-            return this;
-        }
-
-        public LfasrClient.Builder max_alternatives(String max_alternatives) {
-            this.max_alternatives = max_alternatives;
-            return this;
-        }
-
-        public LfasrClient.Builder speaker_number(String speaker_number) {
-            this.speaker_number = speaker_number;
-            return this;
-        }
-
-        public LfasrClient.Builder has_seperate(String has_seperate) {
-            this.has_seperate = has_seperate;
-            return this;
-        }
-
-
-        public LfasrClient.Builder role_type(String role_type) {
-            this.role_type = role_type;
+        public LfasrClient.Builder duration(Long duration) {
+            this.duration = duration;
             return this;
         }
 
@@ -392,8 +563,108 @@ public class LfasrClient {
             return this;
         }
 
+        public LfasrClient.Builder callbackUrl(String callbackUrl) {
+            this.callbackUrl = callbackUrl;
+            return this;
+        }
+
+        public LfasrClient.Builder hotWord(String hotWord) {
+            this.hotWord = hotWord;
+            return this;
+        }
+
+        public LfasrClient.Builder candidate(Short candidate) {
+            this.candidate = candidate;
+            return this;
+        }
+
+        public LfasrClient.Builder roleType(Short roleType) {
+            this.roleType = roleType;
+            return this;
+        }
+
+        public LfasrClient.Builder roleNum(Short roleNum) {
+            this.roleNum = roleNum;
+            return this;
+        }
+
         public LfasrClient.Builder pd(String pd) {
             this.pd = pd;
+            return this;
+        }
+
+        public LfasrClient.Builder audioMode(String audioMode) {
+            this.audioMode = audioMode;
+            return this;
+        }
+
+        public LfasrClient.Builder audioUrl(String audioUrl) {
+            this.audioUrl = audioUrl;
+            return this;
+        }
+
+        public LfasrClient.Builder standardWav(Integer standardWav) {
+            this.standardWav = standardWav;
+            return this;
+        }
+
+        public LfasrClient.Builder languageType(Integer languageType) {
+            this.languageType = languageType;
+            return this;
+        }
+
+        public LfasrClient.Builder trackMode(Short trackMode) {
+            this.trackMode = trackMode;
+            return this;
+        }
+
+        public LfasrClient.Builder transLanguage(String transLanguage) {
+            this.transLanguage = transLanguage;
+            return this;
+        }
+
+        public LfasrClient.Builder transMode(Short transMode) {
+            this.transMode = transMode;
+            return this;
+        }
+
+        public LfasrClient.Builder engSegMax(Integer engSegMax) {
+            this.engSegMax = engSegMax;
+            return this;
+        }
+
+        public LfasrClient.Builder engSegMin(Integer engSegMin) {
+            this.engSegMin = engSegMin;
+            return this;
+        }
+
+        public LfasrClient.Builder engSegWeight(Float engSegWeight) {
+            this.engSegWeight = engSegWeight;
+            return this;
+        }
+
+        public LfasrClient.Builder engSmoothproc(Boolean engSmoothproc) {
+            this.engSmoothproc = engSmoothproc;
+            return this;
+        }
+
+        public LfasrClient.Builder engColloqproc(Boolean engColloqproc) {
+            this.engColloqproc = engColloqproc;
+            return this;
+        }
+
+        public LfasrClient.Builder engVadMdn(Integer engVadMdn) {
+            this.engVadMdn = engVadMdn;
+            return this;
+        }
+
+        public LfasrClient.Builder engVadMargin(Integer engVadMargin) {
+            this.engVadMargin = engVadMargin;
+            return this;
+        }
+
+        public LfasrClient.Builder engRlang(Integer engRlang) {
+            this.engRlang = engRlang;
             return this;
         }
 
@@ -411,5 +682,7 @@ public class LfasrClient {
             }
             return lfasrClient;
         }
+
     }
+
 }
