+++ websdk-java-speech/src/main/java/cn/xfyun/service/lfasr/task/PullResultTask.java
@@ -1,13 +1,12 @@
 package cn.xfyun.service.lfasr.task;
 
-import cn.xfyun.config.LfasrTaskStatusEnum;
-import cn.xfyun.exception.HttpException;
-import cn.xfyun.model.response.lfasr.LfasrMessage;
+import cn.xfyun.model.response.lfasr.LfasrResponse;
 import cn.xfyun.model.sign.LfasrSignature;
+import cn.xfyun.util.StringUtils;
+
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import java.io.IOException;
 import java.security.SignatureException;
 
 /**
@@ -17,30 +16,39 @@ import java.security.SignatureException;
  * @date : 2021年03月15日
  */
 public class PullResultTask extends AbstractTask {
+
     private static final Logger logger = LoggerFactory.getLogger(PullResultTask.class);
 
-    private final String taskId;
+    private static final String serverUrl = "https://raasr.xfyun.cn/v2/api/getResult";
 
-    public PullResultTask(LfasrSignature signature, String taskId) throws SignatureException {
+    private final String orderId;
+
+    private final String resultType;
+
+    public PullResultTask(LfasrSignature signature, String orderId, String resultType) throws SignatureException {
         super(signature);
-        this.param.put("task_id", taskId);
-        this.taskId = taskId;
+        this.param.put("orderId", orderId);
+        if (!StringUtils.isNullOrEmpty(resultType)) {
+            this.param.put("resultType", resultType);
+        }
+        this.orderId = orderId;
+        this.resultType = resultType;
     }
 
     @Override
-    public LfasrMessage call() {
-        LfasrMessage message = new LfasrMessage();
+    public LfasrResponse call() {
+        LfasrResponse response;
         try {
-            message = resolveMessage(this.connector.post("https://raasr.xfyun.cn/api/getResult", this.param));
+            response = resolveResponse(this.connector.post(serverUrl, this.param));
         } catch (Exception e) {
-            logger.warn(getIntro() + " 处理失败", e);
-            message.setOk(LfasrTaskStatusEnum.STATUS_1.getKey());
+            logger.warn("转写结果查询失败，详细信息：【{}】", getIntro(), e);
+            response = LfasrResponse.error("转写结果查询失败");
         }
-        return message;
+        return response;
     }
 
     @Override
     public String getIntro() {
-        return "pull result task: " + this.taskId;
+        return "查询转写结果接口地址：" + serverUrl + "，订单ID: " + this.orderId + ", 结果类型: " + this.resultType;
     }
 }
